Do not rely on dynamic-width `long long` type; use int64_t.

+++ Source/Glide64/DepthBufferRender.cpp
@@ -87,12 +87,12 @@ static int left_z, left_dzdy;
 
 __inline int imul16(int x, int y)        // (x * y) >> 16
 {
-    return (((long long)x) * ((long long)y)) >> 16;
+    return ((int64_t)x * (int64_t)y) >> 16;
 }
 
 __inline int imul14(int x, int y)        // (x * y) >> 14
 {
-    return (((long long)x) * ((long long)y)) >> 14;
+    return ((int64_t)x * (int64_t)y) >> 14;
 }
 
 __inline int idiv16(int x, int y)        // (x << 16) / y
